requirements.txt
requests==2.31.0
fastapi==0.103.2
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.21
alembic==1.12.0
pymysql==1.1.0
openai==0.28.1
boto3==1.17.112
python-decouple==3.8
gunicorn==21.2.0